+++ src/Core/Compiler.cpp
@@ -17,10 +17,10 @@
 
 #include "Core/Compiler.hpp"
 #include "Core/EventLogger.hpp"
+#include "Settings/SettingsManager.hpp"
 #include <QDir>
 #include <QFileInfo>
 #include <QProcess>
-#include <generated/SettingsHelper.hpp>
 
 namespace Core
 {
@@ -68,11 +68,14 @@ void Compiler::start(const QString &tmpFilePath, const QString &sourceFilePath, 
         command = QString("%1 \"%2\" -o \"%3\"")
                       .arg(compileCommand)
                       .arg(QFileInfo(tmpFilePath).canonicalFilePath())
-                      .arg(cppOutputPath(tmpFilePath, sourceFilePath));
+                      .arg(outputPath(tmpFilePath, sourceFilePath, "C++"));
     }
     else if (lang == "Java")
     {
-        command = compileCommand + " \"" + QFileInfo(tmpFilePath).canonicalFilePath() + "\"";
+        command = QString("%1 \"%2\" -d \"%3\"")
+                      .arg(compileCommand)
+                      .arg(QFileInfo(tmpFilePath).canonicalFilePath())
+                      .arg(outputPath(tmpFilePath, sourceFilePath, "Java"));
     }
     else if (lang == "Python")
     {
@@ -97,27 +100,31 @@ bool Compiler::check(const QString &compileCommand)
     // check both "--version" and "-version", "-version" is mainly for Java
 
     checkProcess.start(compileCommand.trimmed().split(' ').front() + " --version");
-    bool finished = checkProcess.waitForFinished(1000);
+    bool finished = checkProcess.waitForFinished(2000);
     if (finished && checkProcess.exitCode() == 0)
         return true;
     checkProcess.kill(); // kill it if it's not finished, no harm if it's finished with non-zero exit code
 
     checkProcess.start(compileCommand.trimmed().split(' ').front() + " -version");
-    finished = checkProcess.waitForFinished(1000);
+    finished = checkProcess.waitForFinished(2000);
 
     LOG_INFO(BOOL_INFO_OF(finished) << INFO_OF(checkProcess.exitCode()));
 
     return finished && checkProcess.exitCode() == 0;
 }
 
-QString Compiler::cppOutputPath(const QString &tmpFilePath, const QString &sourceFilePath)
+QString Compiler::outputPath(const QString &tmpFilePath, const QString &sourceFilePath, const QString &lang)
 {
     QFileInfo fileInfo(sourceFilePath.isEmpty() ? tmpFilePath : sourceFilePath);
-    QString res = fileInfo.dir().filePath(SettingsHelper::getCppExecutableFilePath()
+    QString res = fileInfo.dir().filePath(SettingsManager::get(lang + "/Output Path")
+                                              .toString()
                                               .replace("${filename}", fileInfo.fileName())
                                               .replace("${basename}", fileInfo.completeBaseName())
                                               .replace("${tmpdir}", QFileInfo(tmpFilePath).absolutePath()));
-    QDir().mkpath(QFileInfo(res).absolutePath());
+    if (lang == "C++")
+        QDir().mkpath(QFileInfo(res).absolutePath());
+    else if (lang == "Java")
+        QDir().mkpath(res);
     return res;
 }